feat(Operator): Adds an operator template generation script and its configuration file.
Adds a generator script that generates the basic template for the operator and its corresponding json configuration file.
To use the script, we need to configure the environment variable ASCEND_HOME_PATH to be the installation directory of the CANN package.
Hint according to CANN default path: “/usr/local/Ascend/ascend-toolkit/latest”.