feat(Operator): attempted to write programming paradigm framework, modified generation scripts
Attempted to write programming paradigm framework, some functions are not yet defined, will be developed later.
Fixed the problem that the generated script does not correspond to the local processor version (ascend910b -> ascend910b1).
Simultaneously modified the local processor version in the operator prototype registration. The correctness of this modification will be verified later by comparing it with the new template generated by the modified script.